Split main startup functions
