+++ packages/apps/electron-app/src/main/index.ts
@@ -1,5 +1,4 @@
-import { cli } from "@superego/cli";
-import pkg from "../../package.json" with { type: "json" };
+import startCliProcess from "./startCliProcess.js";
 
 const cliArgs = getCliArgs();
 
@@ -7,7 +6,7 @@ if (cliArgs) {
   void startCliProcess(cliArgs);
 } else {
   const { default: startElectronProcess } =
-    await import("./electronProcess.js");
+    await import("./startElectronProcess.js");
   startElectronProcess();
 }
 
@@ -19,19 +18,4 @@ function getCliArgs(): string[] | null {
     return null;
   }
   return process.argv.slice(separatorIndex + 1);
-}
-
-async function startCliProcess(cliArgs: string[]): Promise<void> {
-  const argv = [process.argv[0] ?? "superego-app", "superego", ...cliArgs];
-
-  try {
-    await cli({ version: pkg.version, argv });
-  } catch (error) {
-    console.error(error instanceof Error ? error.message : String(error));
-    process.exitCode = 1;
-  } finally {
-    const exitCode =
-      process.exitCode === undefined ? 0 : Number(process.exitCode);
-    process.exit(Number.isFinite(exitCode) ? exitCode : 1);
-  }
 }
